Attempt to reduce test flakiness by forcing file sync
Some automated testing has shown this test to fail due to being unable
to read the written version bytes from the file. The root cause would
seem to be something at the OS level, so do a sync_all() to hopefully
ensure that the immediate read after write sees the contents of the
write.