IDL: Remove marker type fields from all types (#4)
Rust marker types (`PhantomData` and `PhantomPinned`) are irrelevant for
TypeScript and they don't have any impact on the memory layout of the
structer. They exist purely for Rust compiler to recognize the ownership
of a generic type which is not a direct memeber of the given struct.

Before this change, `PhantomData` fields were included in IDL and then
TypeScript programs using them were failing to find the type. This
change fixes that just by removing them.