IDL: Expand macros before parsing Rust code (#1)
The IDL parser used to look specifically for attribute macros (e.g. for
accounts). But that way doesn't allow expanding macros before generating
IDL, which prevents writing macros which extend accounts with additional
fields.

For that reason, this change always expands macros before generating IDL
and changes the way the code is parsed:

* Instead for parsing attribute macros (`#[account]`,
  `#[derive(Accounts)]`), it parses expanded `impl` blocks (e.g. `impl
  anchor_lang::Accounts`).
* Explicitly omits expanded code which is not supposed to be included as
  a part of IDL definitions (e.g. modules starting with `__`).