Add a `justfile` for setting up your local machine for tests
I was about to work on a PR to address #2586 and was trying to get my
tests set up so that I could test locally. I found it pretty cumbersome
to copy all of the commands from the ci files in `.github` (maybe
there's a better way? in which case, this isn't necessary), so I threw
them in a `justfile` in case it'd be helpful for others.