Add cargo path update for fish
During tests on a new VM with fish, avm was available immediately after installation of the dependencies and avm itself, without adding the cargo/bin to path or reloading the shell. However, in case someone else installed rust in a way that cargo/bin was not accessible by the system path, maybe it is important to leave in these path update sections. 

My apologies for the delay, and please let me know if there's anything else I should do.